do not process KDL/Joint/JointType
JointType has a Fixed field, which collides with CORBA's IDL "fixed"
keyword. The wrappers already had their own version of the enum, but
KDL's was processed (but not exported) anyways because the JointType
enum appears in joint.hpp. Note that the type was not used as all
KDL types are opaques anyways.

Remove the type explicitely.